baymaxAssistant/app/requirements.txt.txt
streamlit
opencv-python
numpy
scipy
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
